make admin and user page
